CLEAN: Leave one version of coupling browser alone. - The initial function based view and it's succesor in a class base view have been cleaned out. - Column close buttons where they should be. - draw.dt was missing to detect filtering effects on overlay.
