+++ signprot/urls.py
@@ -1,7 +1,7 @@
 from . import views
 from django.urls import path
 from django.views.decorators.cache import cache_page
-from signprot.views import CouplingBrowser, CouplingBrowser2
+from signprot.views import CouplingBrowser
 from contactnetwork.views import PdbTableData
 
 urlpatterns = [
@@ -11,10 +11,8 @@ urlpatterns = [
     path('statistics_venn',  views.GProteinVenn, name='gprotein'),
     path('statistics_tree',  views.GProteinTree, name='gprotein'),
     path('statistics',  views.GProtein, name='gprotein'),
-    path('couplings', views.couplings, name='coupling_browser'),
-    path('couplings1', cache_page(60*60*24*7)(CouplingBrowser.as_view()), name='coupling_browser1'),
-#    path('couplings2', (CouplingBrowser2.as_view()), name='coupling_browser2'),
-    path('couplings2', cache_page(60*60*24*7)(CouplingBrowser2.as_view()), name='coupling_browser2'),
+    path('couplings', (CouplingBrowser.as_view()), name='coupling_browser'),
+#    path('couplings', cache_page(60*60*24*7)(CouplingBrowser.as_view()), name='coupling_browser'),
     path('ginterface/<protein>/', views.Ginterface, name='render'),
     path('ginterface/', views.TargetSelection.as_view(), name='targetselection'),
     path('ajax/barcode/<slug>/<cutoff>/', views.ajaxBarcode, name='ajaxBarcode'),
