+++ signprot/views.py
@@ -134,400 +134,13 @@ class TargetSelection(AbsTargetSelection):
 class CouplingBrowser(TemplateView):
     """
     Class based generic view which serves coupling data between Receptors and G-proteins.
-    Data coming from Guide to Pharmacology, Asuka Inuoue and Michel Bouvier.
-    More data might come later from Roth and Strachan TRUPATH biosensor.
-    :param dataset: ProteinGProteinPair (see build/management/commands/build_g_proteins.py)
-    :return: context
-    """
-    template_name = "signprot/coupling_browser1.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        fam_tab = self.fams_tab()
-
-        context['famtab'] = fam_tab
-
-        return context
-
-    @staticmethod
-    def fams_tab():
-        """
-        This function returns what is needed in the families (maybe others - needs refactoring) tab of the coupling
-        browser.
-        :return: key.value pairs from fd dictionary
-        keys = gene names of proteins
-        values = gpcrDB general values and emax-double-normalized, pec50-double-normalized, emax-mean, pec50-mean
-        emax-sem, pec50sem.
-        """
-        threshold_primary_inoue = 0.5  # -0.1
-        threshold_secondary_inoue = 0.01  # -1
-        threshold_primary_bouvier = 0.5  # -0.1
-        threshold_secondary_bouvier = 0.01  # -1
-        proteins = Protein.objects.filter(sequence_type__slug='wt',
-                                          family__slug__startswith='00',
-                                          species__common_name='Human').all().prefetch_related('family')
-        data = {}
-        class_names = {}
-        family_names = {}
-        for p in proteins:
-            p_class = p.family.slug.split('_')[0]
-            if p_class not in class_names:
-                class_names[p_class] = p.family.parent.parent.parent.name
-                family_names[p_class] = p.family.parent.name
-            p_class_name = class_names[p_class].replace('Class', '').strip()
-            p_family_name = family_names[p_class].replace('receptors', '').strip()
-            p_accession = p.accession
-            p_entryname = p.entry_name
-            data[p.entry_short()] = {'class': p_class_name,
-                                     'family': p_family_name,
-                                     'accession': p_accession,
-                                     'entryname': p_entryname,
-                                     'pretty': p.short(),
-                                     'GuideToPharma': {},
-                                     'Aska': {},
-                                     'Bouvier1': {}}
-
-        distinct_g_families = []
-        distinct_g_subunit_families = {}
-        couplings = ProteinGProteinPair.objects.prefetch_related('protein',
-                                                                 'g_protein_subunit',
-                                                                 'g_protein')
-        for c in couplings:
-            p = c.protein.entry_short()
-
-            # Skip entries without any annotation
-            if p not in data:
-                continue
-
-            s = c.source
-            t = c.transduction
-            emdn = c.emax_dnorm
-            pec50dn = c.pec50_dnorm
-            emmean = c.emax_mean
-            pec50mean = c.pec50_mean
-            emsem = c.emax_sem
-            pec50sem = c.pec50_sem
-            gf = c.g_protein.name
-            gf = gf.replace(" family", "")
-
-            if gf not in distinct_g_families:
-                distinct_g_families.append(gf)
-                distinct_g_subunit_families[gf] = []
-
-            if c.g_protein_subunit:
-                g = c.g_protein_subunit.entry_name
-                g = g.replace("_human", "")
-                if g not in distinct_g_subunit_families[gf]:
-                    distinct_g_subunit_families[gf].append(g)
-                    distinct_g_subunit_families[gf] = sorted(distinct_g_subunit_families[gf])
-
-            if s not in data[p]:
-                data[p][s] = {}
-
-            if gf not in data[p][s]:
-                data[p][s][gf] = {}
-
-            # If transduction (primary, secondary) in Guide To Pharmacology data
-            if t:
-                data[p][s][gf] = t
-
-            # Else t (primary secondary classification) doesn't exist, such as
-            # source [s]  Aska (Inoue) or Bouvier.
-            else:
-                if 'emdn' not in data[p][s][gf]:
-                    data[p][s][gf] = {'emdn': {}, 'max': 0.00, 'maxid': str()}
-                if emdn is None:
-                    data[p][s][gf]['emdn'][g] = round(Decimal(0), 2)
-                    continue
-                data[p][s][gf]['emdn'][g] = round(Decimal(emdn), 2)
-                # get the highest number into 'max'
-                # TODO: Who is max? It will eventually be needed, count on it!
-                if emdn > data[p][s][gf]['max']:
-                    data[p][s][gf]['max'] = round(Decimal(emdn), 2)
-                    data[p][s][gf]['maxid'] = g
-                # pec50dn
-                if 'pec50dn' not in data[p][s][gf]:
-                    data[p][s][gf]['pec50dn'] = {}
-                if pec50dn is None:
-                    data[p][s][gf]['pec50dn'][g] = round(Decimal(0), 2)
-                    continue
-                data[p][s][gf]['pec50dn'][g] = round(Decimal(pec50dn), 2)
-                # emmean
-                if 'emmean' not in data[p][s][gf]:
-                    data[p][s][gf]['emmean'] = {}
-                if emmean is None:
-                    data[p][s][gf]['emmean'][g] = round(Decimal(0), 2)
-                    continue
-                data[p][s][gf]['emmean'][g] = round(Decimal(emmean), 2)
-                # pec50mean
-                if 'pec50mean' not in data[p][s][gf]:
-                    data[p][s][gf]['pec50mean'] = {}
-                if pec50mean is None:
-                    data[p][s][gf]['pec50mean'][g] = round(Decimal(0), 2)
-                    continue
-                data[p][s][gf]['pec50mean'][g] = round(Decimal(pec50mean), 2)
-                # emsem
-                if 'emsem' not in data[p][s][gf]:
-                    data[p][s][gf]['emsem'] = {}
-                if emsem is None:
-                    data[p][s][gf]['emsem'][g] = round(Decimal(0), 2)
-                    continue
-                data[p][s][gf]['emsem'][g] = round(Decimal(emsem), 2)
-                # pec50sem
-                if 'pec50sem' not in data[p][s][gf]:
-                    data[p][s][gf]['pec50sem'] = {}
-                if pec50sem is None:
-                    data[p][s][gf]['pec50sem'][g] = round(Decimal(0), 2)
-                    continue
-                data[p][s][gf]['pec50sem'][g] = round(Decimal(pec50sem), 2)
-
-        fd = {}  # final data
-        distinct_g_families = ['Gs', 'Gi/Go', 'Gq/G11', 'G12/G13']
-        distinct_g_subunit_families = OrderedDict(
-            [('Gs', ['gnas2', 'gnal']),
-             ('Gi/Go', ['gnai1', 'gnai2', 'gnai3', 'gnat1', 'gnat2', 'gnat3', 'gnao', 'gnaz']),
-             ('Gq/G11', ['gnaq', 'gna11', 'gna14', 'gna15']),
-             ('G12/G13', ['gna12', 'gna13'])])
-
-        # This for loop, does the job of putting together
-        # data-sets, GuideToPharma, Asuka's and Bouvier.
-        for p, v in data.items():
-            fd[p] = [v['class'], v['family'], v['accession'], v['entryname'], p, v['pretty']]
-            s = 'GuideToPharma'
-            # MERGED CRITERIA FOR COUPLING
-            # merge primary, secondary, coupling, no coupling classification currently from
-            # three sources, GtP, Inoue, Bouvier. Exact rules being worked on.
-            for gf in distinct_g_families:
-                values = []
-                if 'GuideToPharma' in v and gf in v['GuideToPharma']:
-                    values.append(v['GuideToPharma'][gf])
-                else:
-                    values.append('na gtf')
-                if 'Aska' in v and gf in v['Aska']:
-                    max = v['Aska'][gf]['max']
-                    if max > threshold_primary_inoue:
-                        values.append('primary')
-                    elif max > threshold_secondary_inoue:
-                        values.append('secondary')
-                    elif max == 0:
-                        values.append('NCI')
-                else:
-                    values.append('na inoue')
-                if 'Bouvier1' in v and gf in v['Bouvier1']:
-                    max = v['Bouvier1'][gf]['max']
-                    if max > threshold_primary_bouvier:
-                        values.append('primary')
-                    elif max > threshold_secondary_bouvier:
-                        values.append('secondary')
-                    elif max == 0:
-                        values.append('NCB')
-                else:
-                    values.append('na bouvier')
-                if 'primary' in values:
-                    fd[p].append('primary')
-                elif 'secondary' in values:
-                    fd[p].append('secondary')
-                #                elif 'NCI' or 'NCB' in values:
-                #                    fd[p].append('no coupling')
-                else:
-                    fd[p].append('NA')  # Should this be NA of no-coupling? What's the GtP meaning?
-                # if (len(values) >= 2):
-                #     print(values, gf, p)
-
-            # Loop over GuideToPharma
-            s = 'GuideToPharma'
-            for gf in distinct_g_families:
-                if gf in v[s]:
-                    fd[p].append(v[s][gf])
-                else:
-                    # fd[p].append("NAg")
-                    fd[p].append('')
-
-            # Loop over Aska
-            s = 'Aska'
-            for gf in distinct_g_families:
-                if gf in v[s]:
-                    if v[s][gf]['max'] > threshold_primary_inoue:
-                        fd[p].append("primaryino")
-                    elif v[s][gf]['max'] > threshold_secondary_inoue:
-                        fd[p].append("secondaryino")
-                    else:
-                        fd[p].append("No coup. Ino")
-                else:
-                    fd[p].append('NAino')  # This last empty one means not-available, NOT no-coupling
-
-            # Last bit adds values in subfamilies (i.e. subtypes)
-            for gf, sfs in distinct_g_subunit_families.items():
-                for sf in sfs:
-                    if gf in v[s]:
-                        if sf in v[s][gf]['emdn']:
-                            fd[p].append(v[s][gf]['emdn'][sf])
-                        else:
-                            # fd[p].append("NAi1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("emdn nai")
-                        fd[p].append('')
-
-                    if gf in v[s]:
-                        if sf in v[s][gf]['pec50dn']:
-                            fd[p].append(v[s][gf]['pec50dn'][sf])
-                        else:
-                            # fd[p].append("NAi1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("pec50dn nai")
-                        fd[p].append('')
-
-                    if gf in v[s]:
-                        if sf in v[s][gf]['emmean']:
-                            fd[p].append(v[s][gf]['emmean'][sf])
-                        else:
-                            # fd[p].append("NAi1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("emmean nai")
-                        fd[p].append('')
-
-                    if gf in v[s]:
-                        if sf in v[s][gf]['pec50mean']:
-                            fd[p].append(v[s][gf]['pec50mean'][sf])
-                        else:
-                            # fd[p].append("NAi1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("pec50mean nai")
-                        fd[p].append('')
-
-                    if gf in v[s]:
-                        if sf in v[s][gf]['emsem']:
-                            fd[p].append(v[s][gf]['emsem'][sf])
-                        else:
-                            # fd[p].append("NAi1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("emsem nai")
-                        fd[p].append('')
-
-                    if gf in v[s]:
-                        if sf in v[s][gf]['pec50sem']:
-                            fd[p].append(v[s][gf]['pec50sem'][sf])
-                        else:
-                            # fd[p].append("NAi1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("pec50sem nai")
-                        fd[p].append('')
-
-            # Loop over Bouvier
-            s = 'Bouvier1'
-            for gf in distinct_g_families:
-                if gf in v[s]:
-                    if v[s][gf]['max'] > threshold_primary_bouvier:
-                        fd[p].append("primary")
-                    elif v[s][gf]['max'] > threshold_secondary_bouvier:
-                        fd[p].append("secondary")
-                    else:
-                        fd[p].append("NC")
-                else:
-                    fd[p].append('NA')  # This last empty one means not-available, NOT no-coupling
-
-            # Last bit adds values to subfamilies (i.e. subtypes)
-            for gf, sfs in distinct_g_subunit_families.items():
-                for sf in sfs:
-                    if gf in v[s]:
-                        if sf in v[s][gf]['emdn']:
-                            fd[p].append(v[s][gf]['emdn'][sf])
-                        else:
-                            # fd[p].append("NAb1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("emdn nab")
-                        fd[p].append('')
-
-                    if gf in v[s]:
-                        if sf in v[s][gf]['pec50dn']:
-                            fd[p].append(v[s][gf]['pec50dn'][sf])
-                        else:
-                            # fd[p].append("NAb1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("pec50dn nab")
-                        fd[p].append('')
-
-                    if gf in v[s]:
-                        if sf in v[s][gf]['emmean']:
-                            fd[p].append(v[s][gf]['emmean'][sf])
-                        else:
-                            # fd[p].append("NAb1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("emmean nab")
-                        fd[p].append('')
-
-                    if gf in v[s]:
-                        if sf in v[s][gf]['pec50mean']:
-                            fd[p].append(v[s][gf]['pec50mean'][sf])
-                        else:
-                            # fd[p].append("NAb1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("pec50mean nab")
-                        fd[p].append('')
-
-                    if gf in v[s]:
-                        if sf in v[s][gf]['emsem']:
-                            fd[p].append(v[s][gf]['emsem'][sf])
-                        else:
-                            # fd[p].append("NAb1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("emsem nab")
-                        fd[p].append('')
-
-                    if gf in v[s]:
-                        if sf in v[s][gf]['pec50sem']:
-                            fd[p].append(v[s][gf]['pec50sem'][sf])
-                        else:
-                            # fd[p].append("NAb1")
-                            fd[p].append('')
-                    else:
-                        # fd[p].append("pec50sem nab")
-                        fd[p].append('')
-
-            # max Values for Gs, Gi/Go, Gq/G11, G12/13 and source Inoue
-            s = 'Aska'
-            for gf in distinct_g_families:
-                if gf in v[s]:
-                    # fd[p].append(v[s][gf]['max'] + 100)
-                    fd[p].append(v[s][gf]['max'])
-                else:
-                    # fd[p].append("NAi max")
-                    fd[p].append('')
-
-            # max Values for Gs, Gi/Go, Gq/G11, G12/13 and source Bouvier
-            s = 'Bouvier1'
-            for gf in distinct_g_families:
-                if gf in v[s]:
-                    # fd[p].append(v[s][gf]['max'] + 200)
-                    fd[p].append(v[s][gf]['max'])
-                else:
-                    # fd[p].append("NAb max")
-                    fd[p].append('')
-
-        return fd
-
-
-class CouplingBrowser2(TemplateView):
-    """
-    Class based generic view which serves coupling data between Receptors and G-proteins.
     Data coming from Guide to Pharmacology, Asuka Inuoue and Michel Bouvier at the moment.
     More data might come later from Roth and Strachan TRUPATH biosensor and Neville Lambert.
     :param dataset: ProteinGProteinPair (see build/management/commands/build_g_proteins.py)
     :return: context
     """
 
-    template_name = "signprot/coupling_browser2.html"
+    template_name = "signprot/coupling_browser.html"
 
     @method_decorator(csrf_exempt)
     def get_context_data(self, **kwargs):
@@ -558,7 +171,7 @@ class CouplingBrowser2(TemplateView):
                                           species__common_name='Human').prefetch_related(
             'family',
             'family__parent__parent__parent',
-            'web_links__web_resource'
+            'web_links'
         )
 
         couplings = ProteinGProteinPair.objects.filter(source="GuideToPharma").values_list('protein__entry_name',
@@ -608,13 +221,13 @@ class CouplingBrowser2(TemplateView):
             protein_data[prot.id]['gq11'] = protein_data[prot.id][gprotein_families[2]]
             protein_data[prot.id]['g1213'] = protein_data[prot.id][gprotein_families[3]]
 
-        couplings2 = ProteinGProteinPair.objects.filter(source__in=["Inoue", "Bouvier"])\
-            .filter(g_protein_subunit__family__slug__startswith="100_001").order_by("g_protein_subunit__family__slug", "source")\
+        couplings2 = ProteinGProteinPair.objects.filter(source__in=["Inoue", "Bouvier"]) \
+            .filter(g_protein_subunit__family__slug__startswith="100_001").order_by("g_protein_subunit__family__slug", "source") \
             .prefetch_related('g_protein_subunit__family', 'g_protein')
 
-        coupling_headers = ProteinGProteinPair.objects.filter(source__in=["Inoue", "Bouvier"])\
-            .filter(g_protein_subunit__family__slug__startswith="100_001")\
-            .order_by("g_protein_subunit__family__slug", "source").distinct("g_protein_subunit__family__slug")\
+        coupling_headers = ProteinGProteinPair.objects.filter(source__in=["Inoue", "Bouvier"]) \
+            .filter(g_protein_subunit__family__slug__startswith="100_001") \
+            .order_by("g_protein_subunit__family__slug", "source").distinct("g_protein_subunit__family__slug") \
             .values_list("g_protein_subunit__family__name", "g_protein_subunit__family__parent__name")
 
         coupling_header_names = {}
@@ -815,142 +428,6 @@ def GProteinTree(request, dataset="GuideToPharma"):
 
 def GProteinVenn(request, dataset="GuideToPharma"):
     return GProtein(request, dataset, "venn")
-
-#@cache_page(60*60*24*7)
-def couplings(request):
-    """
-    Presents coupling data between Receptors and G-proteins.
-    Data coming from Guide to Pharmacology, Asuka Inuoue and Michel Bouvier
-    """
-    context = OrderedDict()
-    threshold_primary = -0.1
-    threshold_secondary = -1
-    proteins = Protein.objects.filter(sequence_type__slug='wt', family__slug__startswith='00',
-                                      species__common_name='Human').all().prefetch_related('family')
-    data = {}
-    class_names = {}
-
-    for p in proteins:
-        p_class = p.family.slug.split('_')[0]
-        if p_class not in class_names:
-            class_names[p_class] = re.sub(r'\([^)]*\)', '', p.family.parent.parent.parent.name)
-        p_class_name = class_names[p_class].replace('Class','').strip()
-        data[p.entry_short()] = {'class': p_class_name, 'pretty': p.short()[:15], 'GuideToPharma': {}, 'Aska': {}}
-    distinct_g_families = []
-    distinct_g_subunit_families = {}
-    # distinct_sources = ['GuideToPharma', 'Aska']
-    couplings = ProteinGProteinPair.objects.all().prefetch_related('protein', 'g_protein_subunit', 'g_protein')
-
-    for c in couplings:
-        p = c.protein.entry_short()
-        # Skip entries without any annotation
-        if p not in data:
-            continue
-
-        s = c.source
-        t = c.transduction
-        m = c.log_rai_mean
-        gf = c.g_protein.name
-        # print(gf)
-        gf = gf.replace(" family", "")
-
-        if gf not in distinct_g_families:
-            distinct_g_families.append(gf)
-            distinct_g_subunit_families[gf] = []
-
-        if c.g_protein_subunit:
-            g = c.g_protein_subunit.entry_name
-            g = g.replace("_human", "")
-            # print("g",g)
-            if g not in distinct_g_subunit_families[gf]:
-                distinct_g_subunit_families[gf].append(g)
-                distinct_g_subunit_families[gf] = sorted(distinct_g_subunit_families[gf])
-
-        if s not in data[p]:
-            data[p][s] = {}
-
-        if gf not in data[p][s]:
-            data[p][s][gf] = {}
-
-        # If transduction in GuideToPharma data
-        if t:
-            data[p][s][gf] = t
-        else:
-            if 'subunits' not in data[p][s][gf]:
-                data[p][s][gf] = {'subunits': {}, 'best': -2.00}
-            if m is None:
-                data[p][s][gf]['subunits'][g] = round(Decimal(0), 2)
-                continue
-            data[p][s][gf]['subunits'][g] = round(Decimal(m), 2)
-
-
-            if round(Decimal(m), 2) == -0.00:
-                data[p][s][gf]['subunits'][g] = 0.00
-            # get the lowest number into 'best'
-            if m > data[p][s][gf]['best']:
-                data[p][s][gf]['best'] = round(Decimal(m), 2)
-    fd = {}  # final data
-    distinct_g_families = sorted(distinct_g_families)
-    distinct_g_families = ['Gs', 'Gi/Go', 'Gq/G11', 'G12/G13', ]
-    distinct_g_subunit_families = OrderedDict([('Gs', ['gnas2', 'gnal']), ('Gi/Go', ['gnai1', 'gnai3', 'gnao', 'gnaz']),
-                                               ('Gq/G11', ['gnaq', 'gna14', 'gna15']), ('G12/G13', ['gna12', 'gna13'])])
-
-
-
-    # This for loop, which perhaps should be a function in itself, perhaps an instance of a Couplings class, does
-    # the job of merging together two data-sets, that of the GuideToPharma and Aska's results.
-    for p, v in data.items():
-        fd[p] = [v['class'], p, v['pretty']]
-        s = 'GuideToPharma'
-        # Merge
-        for gf in distinct_g_families:
-            values = []
-            if 'GuideToPharma' in v and gf in v['GuideToPharma']:
-                values.append(v['GuideToPharma'][gf])
-            if 'Aska' in v and gf in v['Aska']:
-                best = v['Aska'][gf]['best']
-                if best > threshold_primary:
-                    values.append('primary')
-                elif best > threshold_secondary:
-                    values.append('secondary')
-            if 'primary' in values:
-                fd[p].append('primary')
-            elif 'secondary' in values:
-                fd[p].append('secondary')
-            else:
-                fd[p].append('')
-        s = 'GuideToPharma'
-        # First loop over GuideToPharma
-        for gf in distinct_g_families:
-            if gf in v[s]:
-                fd[p].append(v[s][gf])
-            else:
-                fd[p].append("")
-        s = 'Aska'
-        for gf in distinct_g_families:
-            if gf in v[s]:
-                if v[s][gf]['best'] > threshold_primary:
-                    fd[p].append("primary")
-                elif v[s][gf]['best'] > threshold_secondary:
-                    fd[p].append("secondary")
-                else:
-                    fd[p].append("No coupling")
-            else:
-                fd[p].append("")
-        for gf, sfs in distinct_g_subunit_families.items():
-            for sf in sfs:
-                if gf in v[s]:
-                    if sf in v[s][gf]['subunits']:
-                        fd[p].append(v[s][gf]['subunits'][sf])
-                    else:
-                        fd[p].append("")
-                else:
-                    fd[p].append("")
-
-    context['data'] = fd
-    context['distinct_gf'] = distinct_g_families
-    context['distinct_sf'] = distinct_g_subunit_families
-    return render(request, 'signprot/coupling_browser.html', context)
 
 #@cache_page(60*60*24*7)
 def familyDetail(request, slug):
